+++ packages/hydrogen/src/foundation/Cache/cache-sub-request.ts
@@ -1,5 +1,5 @@
 import type {QueryKey, CachingStrategy, AllCacheOptions} from '../../types';
-import {getCache} from '../../framework/runtime';
+import {getCache} from '../runtime';
 import {hashKey} from '../../utilities/hash';
 import * as CacheApi from './cache';
 import {CacheSeconds} from './strategies';
