+++ packages/hydrogen/src/hooks/useShopQuery/tests/useShopQuery.test.tsx
@@ -3,7 +3,7 @@ import {useShopQuery} from '../hooks';
 import {mountWithProviders} from '../../../utilities/tests/shopifyMount';
 import {ServerRequestProvider} from '../../../foundation/ServerRequestProvider';
 import {HydrogenRequest} from '../../../foundation/HydrogenRequest/HydrogenRequest.server';
-import {setCache} from '../../../framework/runtime';
+import {setCache} from '../../../foundation/runtime';
 import {InMemoryCache} from '../../../framework/cache/in-memory';
 
 jest.mock('../../../foundation/ssr-interop', () => {
